requirements.txt
black==22.6.0
RPi.bme280==0.2.4
supabase==0.5.8
